refactor: reorganize file structure (pure moves)
- Move message-bus.ts → subscriptions/message-bus.ts
- Move mutation-manager.ts → util/triggers/mutation-manager.ts
- Update all import paths (no logic changes)

This prepares the codebase for modular trigger architecture by
organizing files into logical subdirectories.
